● relocate BlockContext / PersistentBlockContext out of legacy.rs
  Split the block-context pair between its two natural homes:

  - BlockContext (business-layer container, no serde) → new file
    chain_index/types/block_context.rs, outside the db/ subtree.
    Re-exported at chain_index::types::BlockContext.

  - PersistentBlockContext (module-private DB-serde helper) +
    `From<&BlockContext>` / `From<PersistentBlockContext> for BlockContext`
    + `ZainoVersionedSerde for PersistentBlockContext` + round-trip test
    → chain_index/types/db/block.rs (previously a docstring-only stub,
    earmarked for exactly this family of types).

  PBC's visibility tightened from module-private-to-legacy.rs to
  `pub(super)` — sibling consumers in legacy.rs (`IndexedBlock`,
  `BlockHeaderData` encode/decode) still reach it without it leaking into
  the crate's public surface.

  Also:
  - New test `block_context_round_trips_through_persistent` in block.rs
    pins `BlockContext → PersistentBlockContext → BlockContext` as
    identity. Lives beside PBC because PBC is module-private.
  - types/db/block.rs docstring corrected: dropped the stale `BlockIndex`
    bullet (BlockIndex is a business primitive, not a DB type; lives at
    types/primitives.rs), added `PersistentBlockContext` and a note that
    `BlockContext` itself is not a DB type.
