+++ packages/zaino-state/src/chain_index/types.rs
@@ -32,6 +32,7 @@
 //! 2. Used for in-memory operations, conversions, and coordination
 //! 3. Can be changed more freely as they're not persisted
 
+pub mod block_context;
 pub mod db;
 pub mod helpers;
 pub mod primitives;
@@ -40,7 +41,8 @@ pub mod primitives;
 pub use db::legacy::*;
 pub use db::{CommitmentTreeData, CommitmentTreeRoots, CommitmentTreeSizes};
 
-// Re-export business-layer primitives
+// Re-export business-layer primitives and containers
+pub use block_context::BlockContext;
 pub use primitives::BlockIndex;
 
 // Re-export helper types
